requirements.txt
accelerate==1.7.0
albumentations==2.0.8
arxiv==2.2.0
einops==0.8.1
gradio==5.31.0
huggingface_hub==0.32.3
importlib_metadata==8.5.0
lightning==2.5.1.post0
mat4py==0.6.0
matplotlib==3.10.3
mmaction==0.5.0
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.7
model_archiver==1.0.3
model_index==0.1.11
modelindex==0.0.2
numpy==2.2.6
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
prompt_toolkit==3.0.48
pycocoevalcap==1.2
pycocotools==2.0.9
pyperclip==1.9.0
pytest==8.3.4
python-dotenv==1.1.0
pytorch_lightning==2.5.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
rich==14.0.0
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==60.2.0
Sphinx==8.2.3
tabulate==0.9.0
tensorflow==2.19.0
torch==2.5.1
torchmetrics==1.7.0
torchvision==0.20.1
transformers==4.52.4
ts==0.5.1
